+++ app/brochure/views/log-in/index.html
@@ -60,11 +60,21 @@
 
   <p>
     <label for="email">Your email address</label>
-    <input name="email" type="text" value="{{email}}" {{^validemail}} autofocus{{/validemail}} />
+    {{#validemail}}
+    <input name="email" type="text" value="{{email}}" />
+    {{/validemail}}
+    {{^validemail}}
+    <input name="email" type="text" value="{{email}}" autofocus />
+    {{/validemail}}
   </p>
   <p>
     <label for="password">Password <a style="color:inherit;float:right;" href="/log-in/reset">Reset your password</a></label>
-    <input name="password" type="password" {{#validemail}} autofocus{{/validemail}} />
+    {{#validemail}} 
+    <input name="password" type="password" autofocus />
+    {{/validemail}}
+    {{^validemail}} 
+    <input name="password" type="password" />
+    {{/validemail}}
   </p>
   <input type="submit" value="Log in" style="margin-bottom: 8px" />
 </form>
