+++ app/brochure/views/partials/layout-form.html
@@ -5,7 +5,7 @@
 
 <div class="form-container" style="padding-top: 16px;color:;position: relative;">
 	
-        {{{body}}}
+        {{> body}}
         </div>
       </div>
 </body>
